Split user file repository
